--- admin/app/routes/authenticated/complementary-certifications/complementary-certification.js
@@ -1,12 +1,0 @@
-import Route from '@ember/routing/route';
-import { service } from '@ember/service';
-
-export default class ComplementaryCertificationRoute extends Route {
-  @service store;
-
-  model(params) {
-    return this.store.findRecord('complementary-certification', params.complementary_certification_id, {
-      reload: true,
-    });
-  }
-}
