--- admin/app/routes/authenticated/complementary-certifications/complementary-certification/index.js
@@ -1,10 +1,0 @@
-import Route from '@ember/routing/route';
-import { service } from '@ember/service';
-
-export default class IndexRoute extends Route {
-  @service router;
-
-  beforeModel() {
-    this.router.replaceWith('authenticated.complementary-certifications.complementary-certification.details');
-  }
-}
