+++ admin/app/routes/authenticated/complementary-certifications/item/target-profile/new.js
@@ -1,0 +1,22 @@
+import Route from '@ember/routing/route';
+import { service } from '@ember/service';
+
+export default class AttachTargetProfileNewRoute extends Route {
+  @service accessControl;
+
+  beforeModel() {
+    this.accessControl.restrictAccessTo(
+      ['isSuperAdmin'],
+      'authenticated.complementary-certifications.item.target-profile',
+    );
+  }
+
+  model(_) {
+    const complementaryCertification = this.modelFor('authenticated.complementary-certifications.item.target-profile');
+
+    return {
+      complementaryCertification,
+      currentTargetProfile: null,
+    };
+  }
+}
