--- admin/app/templates/authenticated/complementary-certifications/complementary-certification.hbs
@@ -1,7 +1,0 @@
-{{page-title "Certification complémentaire " @model.id " | Pix Admin" replace=true}}
-
-<ComplementaryCertifications::Item::Header @complementaryCertificationLabel={{@model.label}} />
-
-<main class="page-body complementary-certification">
-  {{outlet}}
-</main>
