--- admin/app/templates/authenticated/complementary-certifications/complementary-certification/details.hbs
@@ -1,8 +1,0 @@
-<ComplementaryCertifications::TargetProfiles::Information
-  @complementaryCertification={{@model}}
-  @currentTargetProfile={{this.currentTargetProfile}}
-  @switchTargetProfile={{this.switchTargetProfile}}
-  @switchToggle={{this.isToggleSwitched}}
-/>
-<ComplementaryCertifications::TargetProfiles::BadgesList @currentTargetProfile={{this.currentTargetProfile}} />
-<ComplementaryCertifications::TargetProfiles::History @targetProfilesHistory={{@model.targetProfilesHistory}} />
