+++ admin/app/templates/authenticated/complementary-certifications/item/target-profile.hbs
@@ -1,1 +1,1 @@
-<ComplementaryCertifications::Item::TargetProfile />
+{{outlet}}
