+++ admin/app/templates/authenticated/complementary-certifications/item/target-profile/index.hbs
@@ -1,0 +1,1 @@
+<ComplementaryCertifications::Item::TargetProfile />
