+++ admin/tests/integration/components/complementary-certifications/attach-badges/target-profile-selector/searchbar-test.gjs
