+++ admin/tests/integration/components/complementary-certifications/attach-badges/target-profile-selector/selected-target-profile-test.gjs
